renamed file back to normal
