+++ src/main/java/frc/robot/subsystems/algae/Algae.java
@@ -1,11 +1,10 @@
-package frc.robot.subsystems.algaetemp;
+package frc.robot.subsystems.algae;
 
 import org.littletonrobotics.junction.Logger;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import frc.robot.Constants;
-import frc.robot.subsystems.algae.AlgaeInputsAutoLogged;
 
 /**
  * Algae Subsystem
