+++ src/main/java/frc/robot/subsystems/algae/AlgaeReal.java
@@ -1,4 +1,4 @@
-package frc.robot.subsystems.algaetemp;
+package frc.robot.subsystems.algae;
 
 import com.revrobotics.spark.SparkBase.PersistMode;
 import com.revrobotics.spark.SparkBase.ResetMode;
